Added new class to handle the picture library
Wrote and passed a test to check that the initial tile generation checks
all directories within the root and generates the correct number of
tiles.

Began to tie together prepping an image to a cropped and scaled
BufferedImage, and generating the color array for a Tile, and then
bringing them together into a single entity.